+++ portfolio-frontend/package.json
@@ -8,7 +8,9 @@
     "axios": "^0.19.0",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
-    "react-scripts": "3.1.1"
+    "react-router-dom": "^5.1.0",
+    "react-scripts": "3.1.1",
+    "yarn": "^1.17.3"
   },
   "scripts": {
     "start": "react-scripts start",
