+++ portfolio-frontend/src/App.css
@@ -2,21 +2,6 @@
   text-align: center;
 }
 
-.App-Background {
-  background-image:url(./images/backImg.jpg);
-  overflow:hidden;
-  position:relative;
-  width:100%;
-  max-width:2560px;
-  height:100vh;
-  min-height:740px;
-  margin:0 auto;
-  background-position:50% 80%;
-  background-repeat:no-repeat;
-  background-size:cover;
-  filter: brightness(70%);
-}
-
 .Title_Text {
   position: absolute;
   top: 30%;
