+++ portfolio-frontend/src/components/Footer.js
