+++ portfolio-frontend/src/pages/Home.css
@@ -8,3 +8,22 @@
     border: 3px solid;
     margin: auto;
 }
+.Title_Text {
+    color: yellow;
+    font-family: sans-serif
+
+}
+.App-Background {
+    background-image:url(../images/backImg.jpg);
+    overflow:hidden;
+    position:relative;
+    width:100%;
+    max-width:2560px;
+    height:100vh;
+    min-height:740px;
+    margin:0 auto;
+    background-position:50% 80%;
+    background-repeat:no-repeat;
+    background-size:cover;
+    filter: brightness(70%);
+  }
