+++ portfolio-frontend/src/pages/Introduction.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import IntroductionPanel from '../components/IntroductionPanel';
-import IntroductionAdd from '../components/IntroductionAdd';
+import IntroductionPanel from '../components/introductions/IntroductionPanel';
+import IntroductionAdd from '../components/introductions/IntroductionAdd';
+import DialogFull from '../components/DialogFull';
 import * as service from '../service/introduction';
 
 
@@ -27,23 +28,26 @@ class Introduction extends React.Component {
     render() {
         return (
             <div>
-                {
-                    this.state.introductions ?
-                        this.state.introductions.map(c => {
-                            return (<IntroductionPanel
-                                key={c.idx}
-                                idx={c.idx}
-                                title={c.title}
-                                growth={c.growth}
-                                reason={c.reason}
-                                strength={c.strength}
-                                weakness={c.weakness}
-                                aspiration={c.aspiration}
-                            />
-                            )
-                        }) : ""
-                }
-                <IntroductionAdd />
+                <DialogFull />
+                <div>
+                    {
+                        this.state.introductions ?
+                            this.state.introductions.map(c => {
+                                return (<IntroductionPanel
+                                    key={c.idx}
+                                    idx={c.idx}
+                                    title={c.title}
+                                    growth={c.growth}
+                                    reason={c.reason}
+                                    strength={c.strength}
+                                    weakness={c.weakness}
+                                    aspiration={c.aspiration}
+                                />
+                                )
+                            }) : ""
+                    }
+                    <IntroductionAdd />
+                </div>
             </div>
         );
     }
